requirements.txt
numpy~=2.2.1
fastapi~=0.115.6
torch~=2.5.1
tqdm~=4.67.1
python-dotenv~=1.0.1
pydantic~=2.10.5
uvicorn~=0.34.0
faiss-cpu
nltk
pandas
